+++ src/js/lang/default.js
@@ -1,5 +1,6 @@
 export default {
     collapseAll: 'Collapse all',
+    collapseNode: 'Collapse',
     expandAll: 'Expand all',
-    toggle: 'Toggle',
+    expandNode: 'Expand',
 };