+++ src/js/shapes/languageShape.js
@@ -2,6 +2,7 @@ import PropTypes from 'prop-types';
 
 export default PropTypes.shape({
     collapseAll: PropTypes.string.isRequired,
+    collapseNode: PropTypes.string.isRequired,
     expandAll: PropTypes.string.isRequired,
-    toggle: PropTypes.string.isRequired,
+    expandNode: PropTypes.string.isRequired,
 });