+++ test/CheckboxTree.jsx
@@ -292,16 +292,17 @@ describe('<CheckboxTree />', () => {
             render(
                 <CheckboxTree
                     lang={{
-                        expandAll: 'Expand it',
-                        collapseAll: 'Collapse it',
-                        toggle: 'Toggle it',
+                        expandAll: 'Expand all of it',
+                        expandNode: 'Expand it',
+                        collapseAll: 'Collapse all of it',
+                        collapseNode: 'Collapse it',
                     }}
                     nodes={[]}
                     showExpandAll
                 />,
             );
 
-            assert.isNotNull(screen.queryByLabelText('Expand it'));
+            assert.isNotNull(screen.queryByLabelText('Expand all of it'));
         });
     });
 
@@ -937,7 +938,7 @@ describe('<CheckboxTree />', () => {
             );
 
             const user = userEvent.setup();
-            await user.click(screen.getByLabelText('Toggle'));
+            await user.click(screen.getByLabelText('Expand'));
 
             assert.deepEqual(actualExpanded, ['jupiter']);
         });
@@ -964,7 +965,7 @@ describe('<CheckboxTree />', () => {
             );
 
             const user = userEvent.setup();
-            await user.click(screen.getByLabelText('Toggle'));
+            await user.click(screen.getByLabelText('Expand'));
 
             assert.equal(actualNode.value, 'jupiter');
         });
@@ -1051,7 +1052,7 @@ describe('<CheckboxTree />', () => {
             assert.deepEqual(getNodeMetadata(clickNode), expectedLeafMetadata);
 
             // onExpand
-            await user.click(screen.getByLabelText('Toggle'));
+            await user.click(screen.getByLabelText('Collapse'));
             assert.deepEqual(getNodeMetadata(expandNode), expectedParentMetadata);
         });
     });